fix the placeholder and moved topic handler to utils
